analysis/requirements.txt
fastapi
uvicorn
numpy
opencv-python-headless
requests
pydantic
sentinelhub
shapely
python-dotenv
pytest
httpx
